--- src/types/index.d.ts
@@ -1,9 +1,0 @@
-export namespace Types {
-
-    export { Tools } from './tools'
-
-    export { JSON } from './json'
-
-    export { Shader } from './shader'
-
-}
