--- src/types/shader.d.ts
@@ -1,5 +1,0 @@
-export type Shader = {
-    uniforms: { [uniform: string]: IUniform<any> };
-    vertex: string,
-    fragment: string
-}
