--- src/types/tools.d.ts
@@ -1,9 +1,0 @@
-import { Tools } from '../tools'
-
-export namespace Tools {
-
-    export type Directions = typeof Tools.Directions[number]
-
-    export type Uniforms = { [uniform: string]: IUniform<any> }
-
-}
